feat: Add a hook to retrieve data from Appwrite
