+++ lib/appwrite.js
@@ -101,7 +101,11 @@ export async function getCurrentUser() {
 
 export async function getAllPosts() {
   try {
-    const posts = await databases.listDocuments(databaseId, videoCollectionId);
+    const posts = await databases.listDocuments(
+      config.databaseId,
+      config.videoCollectionId
+    );
+    console.log("posts", posts);
     return posts.documents;
   } catch (error) {}
 }
